requirements.txt
Flask[async]~=2.0.0
pyjwt~=2.3.0
python-dotenv~=0.19.0
Flask-Cors~=3.0.10
asyncpg~=0.25.0
passlib~=1.7.4